+++ cmd/index/index.go
@@ -30,13 +30,8 @@ import (
 // indexCmd represents the index command
 var indexCmd = &cobra.Command{
 	Use:   "index",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Short: "Index commands",
+	Long:  `Index commands`,
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Println("index called")
 	},
